backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.12.1  # Database migrations

# MQTT
paho-mqtt==2.0.0

# Authentication & Security
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator>=1.1.0,<2.1

# Environment variables
python-dotenv==1.0.0

# Data validation & serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSockets
websockets==12.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0  # For health checks 